+++ python/GafferUI/PlugLayout.py
@@ -761,7 +761,10 @@ class _CollapsibleLayout( _Layout ) :
 
 			collapsible.getChild().update( subsection )
 
-			collapsible.setVisible( any ( [ w.getVisible() for w in subsection.widgets ] ) )
+			collapsible.setVisible(
+				any( [ w.getVisible() for w in subsection.widgets ] ) or
+				any( [ w.getVisible() for w in collapsible.getChild().__collapsibles.values() ] )
+			)
 
 			collapsible.getCornerWidget().setText(
 				"<small>" + "&nbsp;( " + subsection.summary + " )</small>" if subsection.summary else ""