+++ Changes.md
@@ -9,8 +9,8 @@ Improvements
 Fixes
 -----
 
+- PlugLayout : Fixed visibility of collapsible layouts with nesting (#4694).
 - Image Node Mix : Fixed incorrect results outside mask data window, and incorrect results when changing inputs.
-
 
 0.61.11.0 (relative to 0.61.10.0)
 =========